+++ lib/standalone.js
@@ -11,24 +11,23 @@ const fs = require('fs');
 const getFormatterOptionsText = require('./utils/getFormatterOptionsText');
 const globby = require('globby');
 const hash = require('./utils/hash');
-const invalidScopeDisables = require('./invalidScopeDisables');
-const needlessDisables = require('./needlessDisables');
 const NoFilesFoundError = require('./utils/noFilesFoundError');
 const path = require('path');
 const pkg = require('../package.json');
+const prepareReturnValue = require('./prepareReturnValue');
 const { default: ignore } = require('ignore');
 const DEFAULT_IGNORE_FILENAME = '.stylelintignore';
 const FILE_NOT_FOUND_ERROR_CODE = 'ENOENT';
 const ALWAYS_IGNORED_GLOBS = ['**/node_modules/**'];
 const writeFileAtomic = require('write-file-atomic');
 
-/** @typedef {import('stylelint').StylelintStandaloneOptions} StylelintOptions */
+/** @typedef {import('stylelint').StylelintStandaloneOptions} StylelintStandaloneOptions */
 /** @typedef {import('stylelint').StylelintStandaloneReturnValue} StylelintStandaloneReturnValue */
 /** @typedef {import('stylelint').StylelintResult} StylelintResult */
 /** @typedef {import('stylelint').Formatter} Formatter */
 
 /**
- * @param {StylelintOptions} options
+ * @param {StylelintStandaloneOptions} options
  * @returns {Promise<StylelintStandaloneReturnValue>}
  */
 module.exports = function (options) {
@@ -47,7 +46,6 @@ module.exports = function (options) {
 	const ignoreDisables = options.ignoreDisables;
 	const reportNeedlessDisables = options.reportNeedlessDisables;
 	const reportInvalidScopeDisables = options.reportInvalidScopeDisables;
-	const maxWarnings = options.maxWarnings;
 	const syntax = options.syntax;
 	const allowEmptyInput = options.allowEmptyInput || false;
 	const useCache = options.cache || false;
@@ -122,7 +120,7 @@ module.exports = function (options) {
 			absoluteCodeFilename &&
 			!filterFilePaths(ignorer, [path.relative(process.cwd(), absoluteCodeFilename)]).length
 		) {
-			return Promise.resolve(prepareReturnValue([]));
+			return Promise.resolve(prepareReturnValue([], options, formatterFunction));
 		}
 
 		return stylelint
@@ -157,7 +155,7 @@ module.exports = function (options) {
 			.catch(_.partial(handleError, stylelint))
 			.then((stylelintResult) => {
 				const postcssResult = stylelintResult._postcssResult;
-				const returnValue = prepareReturnValue([stylelintResult]);
+				const returnValue = prepareReturnValue([stylelintResult], options, formatterFunction);
 
 				if (options.fix && postcssResult && !postcssResult.stylelint.ignored) {
 					if (!postcssResult.stylelint.disableWritingFix) {
@@ -282,50 +280,12 @@ module.exports = function (options) {
 				fileCache.reconcile();
 			}
 
-			return prepareReturnValue(stylelintResults);
+			const rtn = prepareReturnValue(stylelintResults, options, formatterFunction);
+
+			debug(`Linting complete in ${Date.now() - startTime}ms`);
+
+			return rtn;
 		});
-
-	/**
-	 * @param {StylelintResult[]} stylelintResults
-	 * @returns {StylelintStandaloneReturnValue}
-	 */
-	function prepareReturnValue(stylelintResults) {
-		const errored = stylelintResults.some(
-			(result) => result.errored || result.parseErrors.length > 0,
-		);
-
-		/** @type {StylelintStandaloneReturnValue} */
-		const returnValue = {
-			errored,
-			results: [],
-			output: '',
-		};
-
-		if (reportNeedlessDisables) {
-			returnValue.needlessDisables = needlessDisables(stylelintResults);
-		}
-
-		if (reportInvalidScopeDisables) {
-			returnValue.invalidScopeDisables = invalidScopeDisables(stylelintResults);
-		}
-
-		if (maxWarnings !== undefined) {
-			const foundWarnings = stylelintResults.reduce((count, file) => {
-				return count + file.warnings.length;
-			}, 0);
-
-			if (foundWarnings > maxWarnings) {
-				returnValue.maxWarningsExceeded = { maxWarnings, foundWarnings };
-			}
-		}
-
-		returnValue.output = formatterFunction(stylelintResults, returnValue);
-		returnValue.results = stylelintResults;
-
-		debug(`Linting complete in ${Date.now() - startTime}ms`);
-
-		return returnValue;
-	}
 };
 
 /**
